refactor: view에 도메인 변환 및 DTO 변환 책임 추가
