+++ src/main/java/janggi/JanggiGame.java
@@ -7,14 +7,9 @@ import janggi.domain.board.BoardInitializer;
 import janggi.domain.board.ElephantSetting;
 import janggi.domain.board.StandardBoardInitializer;
 import janggi.domain.piece.Camp;
-import janggi.domain.piece.Piece;
 import janggi.dto.CampDto;
-import janggi.dto.PiecePositionDto;
-import janggi.exception.ExceptionMessage;
 import janggi.view.InputView;
 import janggi.view.OutputView;
-import java.util.List;
-import java.util.Map;
 import java.util.function.Supplier;
 
 public class JanggiGame {
@@ -29,7 +24,7 @@ public class JanggiGame {
 
     public void run() {
         Board board = createBoard();
-        outputView.printBoard(toPiecePositions(board.getBoard()));
+        outputView.printBoard(board.getBoard());
         play(board);
     }
 
@@ -46,45 +41,26 @@ public class JanggiGame {
         return new Board(initializer);
     }
 
-    private List<PiecePositionDto> toPiecePositions(Map<Position, Piece> boardState) {
-        return boardState.entrySet().stream()
-                .map(entry -> PiecePositionDto.of(entry.getKey(), entry.getValue()))
-                .toList();
-    }
-
     private void play(Board board) {
         Turn turn = new Turn();
         while (true) {
             retryOnInvalidInput(() -> playTurn(board, turn));
-            outputView.printBoard(toPiecePositions(board.getBoard()));
+            outputView.printBoard(board.getBoard());
         }
     }
 
     private void playTurn(Board board, Turn turn) {
         Camp camp = turn.currentTurn();
-        Position source = readSource(board, camp);
-        Position destination = retryOnInvalidInput(() -> toPosition(inputView.readDestination()));
+        Position source = retryOnInvalidInput(() -> readSource(board, camp));
+        Position destination = retryOnInvalidInput(inputView::readDestination);
         board.movePiece(source, destination, camp);
         turn.finishTurn();
     }
 
     private Position readSource(Board board, Camp camp) {
-        return retryOnInvalidInput(() -> {
-            Position source = toPosition(inputView.readSource(CampDto.from(camp)));
-            board.validateCampTurn(source, camp);
-            return source;
-        });
-    }
-
-    private Position toPosition(List<Integer> rawPosition) {
-        validatePositionSize(rawPosition);
-        return new Position(rawPosition.get(0), rawPosition.get(1));
-    }
-
-    private void validatePositionSize(List<Integer> rawPosition) {
-        if (rawPosition.size() != 2) {
-            throw new IllegalArgumentException(ExceptionMessage.INVALID_INPUT_FORMAT.getMessage());
-        }
+        Position source = inputView.readSource(CampDto.from(camp));
+        board.validateCampTurn(source, camp);
+        return source;
     }
 
     private <T> T retryOnInvalidInput(Supplier<T> input) {
